HUB ETL BASE PDFs/Parser_PDF_Monitoramentos_Hospitalares/requirements.txt
pandas
tabula-py
openpyxl 
tkcalendar
tkinter